Set up the MPU over a TWI interface.
An instance of the nrf library TWI sensor is set to work with the MPU
hardware. The necessary `sdk_config.h` flags were set as well. The
TWI instance is not available outside of the new mpu.h library. The TWI
configuration is kept inside the mpu.h for the moment. MPU settings are
not user-configurable either. A future edition might make it more
modular.

The MPU's internal clock is used to drive 200Hz interrupts, each time a
new measurement is available. This interrupt causes non-blocking reads.
A sensor fusion "placeholder" is invoked on each interrupt.

The magnetometer is read at 100Hz. Each time the magnetometer is
available, more advanced sensor fusion is used.

The nrf TWI sensor libraries use three levels of abstraction:

 - the `nrfx_twi` driver (regretably it is set to the older `nrf_drv_twi`)
 namespace for the moment;
 - the `twi_mngr` for scheduling communication;
 - the `twi_sensor` for abstracting out simple sensors.

For the initialization of the MPU we use nrfx_twi in blocking mode. For
everything else we use the twi_sensor in async mode.