Enable the MadgwickAHRS.h and permit MPU9255.
The sensor fusion library is now enabled. Fusion is running at each
interrupt. Whenever the slower magnetometer reports a result, the more
advanced 9DOF fusion is used.

A side issue noticed in testing:

The MPU9255 is frequently sold as a substitute for MPU9250. The only
difference that concerns the current code base is the different
`WHO_AM_I` code. We now permit both 9250 and 9255.