Use orientation estimate to get kinematic accel.
The kinematic acceleration can be deduced from the current estimate of
the orientation `q` and the current measured acceleration `a` in the
sensor frame:

lab frame a = q^-1 * a * q

kinematic acceleration = lab frame a - g

where g = (0,0,1) is the gravitational acceleration.

The kinematice acceleration (in the lab frame) is provided as a BLE
characteristic as well.

The characteristics being transferred result in a buffer saturation. For
the moment such events that trigger the "no resources" error are just
skipped (leading to a skipped notification a couple of times per minute).