Enable magnetometer in sensor fusion.
A simple magnetometer calibration routine was implemented.

The B field measured by the magnetometer should be of constant magnitude
(it is only the locally homogeneous field of Earth). Therefore, as the
magnetometer rotates, the vectors it measures should form a spherical
shell of points. "Hard iron" errors cause the center of the sphere to
not be at the origin. "Soft iron" errors cause the sphere to become an
ellipse.

By detecting rotations, we measure the points where the ellipse crosses
the x, y, and z axis, and use this to calibrate the measurements.
Initial estimates for these errors are provided as well. Fusing the
initial estimates with current updates is poorly implemented.

We assume that the primary axes of the ellipse are parallel to x, y, and
z.

The corrected magnetometer measurements are sent to the sensor fusion
algorithm.

Simple drift estimation is provided for the gyroscope as well.