Change LF clock source to SYNTH.
The internal RC oscillator could cause connection issues if not properly
calibrated. Instead, we just use the HF clock to synthesize the LF one.

This results in much higher current consumption as we can not turn off
the HF oscillator. Given that we are already keeping the FPU on all the
time, this might be a negligible contribution though (needs to be
checked).