Implement simple multilateration based on signal strength
A buffer of fixed size (currently 100) is filled with measurements of beacon
signal strength (the beacons announce their location) and used to calculate
location.

Description of the multilateration algorithm:

Multilateration can be done based on (noisy) distance measurements to a
set of beacons with known positions. The multilateration is performed by
attempting to solve the system of nonlinear equations:

(x[i]-X)^2 + (y[i]-Y)^2 = r[i]^2

where x[i] and y[i] are the position of the i-th beacon and r[i] is
the approximate distance to that beacon. X and Y are the unknown
position of the receiver.

The i=0 equation is substracted from the other equations in order to
linearize the system, at which linear least squares is performed on the
data.

The list of beacons is stored in a ring buffer, in order to permit easy
on-the-fly updates with newly discovered beacons.

Thread locking is not implemented yet, but it is marked with TODO flags.